refactor: 패키지 이동
- Splitter 클래스를 역할에 따라 regex 패키지에서 splitter 패키지로 이동한다.
